[feat/3] refactor: articleList의 clickEvent, pageEvent, html 생성 함수 분리
